Extension of ls1 by the so called MmspdWriter. This class writes out a *.mmspd-file that is used by MegaMol. A detailed specification of the *.mmspd-file format is given in the MegaMol documentation (see website). The file format implemented contains the trajectories and accounts for up to 5 components distinguished by colour and radius. The latter one is scaled by the Lennard-Jones sigma. A single particle is represented by a sphere which means that solely 1 center LJ particles should be considered for graphical representation.
Since both the total number of components and the LJ sigma are to be known in this file the appropriate methods were implemented in the class Domain 
(a pointer to Domain is an input parameter of all output writers). 
They are called unsigned Domain::getNumberOfComponents()' and 'Domain::getSigma(unsigned cid, unsigned nthSigma)'. 
The falgs in the config file activating the MmspdWriter are: "MmspdWriter" outputFrequency outputPathAndPrefix


git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/branches/density-profiles@1046 a63bd714-7e14-4b5e-94e7-302c8c8ff188