The scenario generator mkSD (sessile drop): The generator is creating a start configuration for ls1 MarDyn that is to be used for simulating sessile drops. So far, the fluid used is a 1CLJTS. The same holds for the wall. The usaged of the application is displayed if it is called without any arguments. @brief: The application is called mkSD. It consists of the following parts:
(i)	 main.cpp => the main() method
(ii) 	 ConfigWriter => instance of this class is wirting the config file in the 'old style'.
(iii)	 PhaseSpaceWriter => instance of this class is writing the input file *.inp
(iv) 	 GlobalStartGeometry => instance of this class determines the box lengths and the major lengths such as offsets etc.
	 Moreover it contributes to the calculation of the positions of the fluid molecules.
(v)	 Component => instances of this class contain the component related information such as sigma, epsilon, ...
(vi)	 Molecule => instance of this class is used to determine information concerning a single particle such as a particles position or velocity.
	 Owed to the way of initializing the position of the fluid particles the instance Molecule DOES NOT calculate the position and velocities, resp.,  of the fluid particles
(vii)	 RandomNumber => instance of this class is generating random numbers




git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/branches/density-profiles@1087 a63bd714-7e14-4b5e-94e7-302c8c8ff188